backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.50.2
pydantic==2.9.2
starlette==0.38.5
